TECH_DOCS/FINANCE_ECONOMICS/CRYPTOECONOMICS/NOMIX/requirements.txt
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.2
requests==2.26.0
scikit-learn==0.24.2
jupyter==1.0.0
pyppeteer==0.2.6
playwright==1.17.0
python-dotenv==0.19.2
openai==0.10.2